requirements.txt
openeo
requests
shapely
pandas
geopandas
numpy
rasterio
pyrosm
sentinelhub
python-dotenv
imageio